requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pdfplumber>=0.10.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.2
scikit-learn>=1.3.0
anthropic>=0.20.0
python-multipart>=0.0.6
